Allow desktop icons to be passed attributes in object
I'm adding this on my own fork so I can pass in files to programs. For example having a general Notepad function but allowing instances (e.g. myNote.txt) to be passed in. 

I thought it might be useful for other programs as well. For example with paint you may want to have an icon representing a specific picture which was painted.

Let me know what you think.