98.js.org
Freenom took away my domain without notice, and either put ads on it or sold it to an anonymous third party who put ads on it. 98desktop.ml is now parked.

So I'm switching to https://98.js.org, which is actually a nicer name anyways: https://github.com/js-org/dns.js.org/pull/1773

As a bonus, this should automatically enable HTTPS via CloudFlare, which I was planning to do to fix Sound Recorder not working in Chrome.