Use SVG filter effects for selected icons
This was an ordeal. Enables support in Edge for the selected icon effect, which previously rendered in a rather broken looking way.
Edge doesn't support clip-path. Or @supports. Or any number of other things that might let us do this somewhat reasonably.
Also it was a pain to use in the debugging process and just as a browser (its basic text selection and drag and drop facilities suck).

In Firefox this causes clipping of folder icons (which are the widest) and blurring https://bugzilla.mozilla.org/show_bug.cgi?id=852100