Use CSS backgrounds for Sound Recorder buttons
And render the disabled state per the entire sprite sheet at once. This might not totally be the way to go because if this was generalized to more of the operating system (disabled buttons look similar in many places), it might run into sprites bleeding into each other. And already I have to use an extra element for the background to clip it to a smaller region. Well, maybe I could use background-clip: content-box;