Animate wallpaper with feDisplacementMap
I was originally trying to implement the science wallpaper,
but it didn't work in chrome. Instead I got an interesting animation effect for the wallpaper,
where it seems to use the SourceGraphic as the displacement map when applying the filter to an HTML element.
Displacing by the input gives a nice flow to the animation, but there's two problems:
1. The animation breaks down (and gets laggy) over time, and can't loop nicely.
2. This only works in Chrome, due to the buggy behavior. Could I even get it to work cross browser??