Use mask-image for icon selection effect
This introduced a flickering that I had to fix by making it not remove and immediately add classes, which is somewhat worrying.
I mean it's not an ideal pattern, but I'd like to be able to do it if it's easier, you know?

Update: yeah, this makes it flicker also when double clicking...