Refactor $Iframe -> enhance_iframe
The biggest part of this is changing the code to let apps handle mouseup events outside their frames to register event listeners only during a gesture,
in order to eliminate the need for a destroy() method. Theoretically it should be garbage collected (although I haven't vetted all the code; there's probably other leaks. Anyways this code seems to work about as well as it did before...