Improve performance?
It's really getting bad over time, I'm guessing due to memory leaks.
It slows down to a crawl, even when just left open in the background.

This may help a little, but not enough, and it may cause problems if I'm caching this when trying to skin elements that have a dynamic state.