Fix selecting items by clicking in Firefox
Closes https://github.com/1j01/98/issues/92

This applies to the file explorer, and not the desktop,
due to the app using iframes.

This was broken when introducing nested iframes to the file explorer in:
https://github.com/1j01/98/commit/0c7a9e7d3c38c09a953c404

`instanceof` is inconsistent between Chrome and Firefox when an element
is created in one document and then appended to a different document:
https://jsfiddle.net/1j01/m8t19oyv/
https://bugzilla.mozilla.org/show_bug.cgi?id=1470017

In Firefox (version 127.0), the element is considered an instance of
the `Element` class of the document that it's appended to.

In Chrome (version 126.0.6478.127), the element is considered
an instance of the original scope's `Element`.